Mover modal nueva entidad a componente aparte
